backend/requirements.txt
boto3==1.26.0
psycopg2-binary==2.9.5
requests==2.28.1
python-dotenv==0.21.0
pandas==1.5.1
